requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
Django==3.1
django-admin-honeypot==1.1.0
django-filter==2.4.0
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
pycryptodome==3.10.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.6
whitenoise==5.2.0
